@@ -54,51 +55,44 @@
   },
   "main": "./dist/server.js",
   "peerDependencies": {
-    "@angular/core": ">=5.0.0 <6.0.0",
-    "@angular/router": ">=5.0.0 <6.0.0"
+    "@angular/core": ">=6.0.0 <7.0.0",
+    "@angular/router": ">=6.0.0 <7.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/platform-server": "^5.2.9",
-    "@angular/router": "^5.2.9",
-    "@angular/service-worker": "^5.2.9",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/entity": "^5.2.0",
-    "@ngrx/router-store": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "@ngrx/store-devtools": "^5.1.0",
-    "@nguniversal/express-engine": "^5.0.0-beta.5",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
-    "core-js": "^2.4.1",
-    "express": "^4.16.2",
-    "hammerjs": "^2.0.8",
-    "rxjs": "^5.5.7",
-    "trailpack-express": "^2.0.6",
-    "trailpack-proxy-router": "^2.1.1",
-    "trailpack-repl": "^2.1.0",
-    "trailpack-router": "^2.0.0",
-    "trails": "^2.0.0",
-    "winston": "^2.2",
-    "zone.js": "^0.8.14",
-    "@angular/cli": "1.6.7",
-    "@angular/compiler-cli": "^5.2.9",
-    "@angular/language-service": "^5.2.9",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/animations": "^6.0.0",
+    "@angular/cdk": "^6.0.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/platform-server": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "@angular/service-worker": "^6.0.0",
+    "@ngrx/effects": "^6.0.0-beta.1",
+    "@ngrx/entity": "^6.0.0-beta.1",
+    "@ngrx/router-store": "^6.0.0-beta.1",
+    "@ngrx/store": "^6.0.0-beta.1",
+    "@ngrx/store-devtools": "^6.0.0-beta.1",
+    "@nguniversal/express-engine": "^6.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~9.4.0",
     "@types/winston": "^2.2.20-alpha",
     "codelyzer": "^4.0.1",
+    "core-js": "^2.4.1",
     "eslint": "^4.17.0",
     "eslint-config-trails": "^3.1.0",
+    "express": "^4.16.2",
+    "hammerjs": "^2.0.8",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
